Add communication command handlers for discovery, connection, sending, status, disconnection, and pin configuration
