+++ include/core/cli-handlers/comm-cli-handler.h
@@ -1,0 +1,8 @@
+#pragma once
+
+void handle_comm_discovery(int argc, char **argv);
+void handle_comm_connect(int argc, char **argv);
+void handle_comm_send(int argc, char **argv);
+void handle_comm_status(int argc, char **argv);
+void handle_comm_disconnect(int argc, char **argv);
+void handle_comm_setpins(int argc, char **argv);
